requirements.txt
alabaster==0.7.12
attrs==21.2.0
Babel==2.9.1
black==21.10b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coverage==6.1.1
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==3.0.2
loguru==0.5.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
regex==2021.11.2
requests==2.26.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
